--- geode-deployment-legacy/src/main/resources/META-INF/services/org.apache.geode.deployment.internal.JarDeploymentService
@@ -1,1 +1,0 @@
-org.apache.geode.internal.deployment.legacy.LegacyJarDeploymentService
