+++ geode-deployment/build.gradle
@@ -14,19 +14,3 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-apply from: "${rootDir}/${scriptDir}/standard-subproject-configuration.gradle"
-
-apply plugin: 'war'
-
-
-
-dependencies {
-  compile(platform(project(':boms:geode-all-bom')))
-  implementation('javax.servlet:javax.servlet-api')
-  implementation(project(':geode-deployment:geode-deployment-legacy'))
-}
-
-war {
-    version = ''
-}
