+++ geode-deployment/geode-deployment-legacy/build.gradle
@@ -15,18 +15,23 @@
  * limitations under the License.
  */
 
-apply from: "${rootDir}/${scriptDir}/standard-subproject-configuration.gradle"
+apply from: "${project.projectDir}/../../gradle/standard-subproject-configuration.gradle"
 
-apply plugin: 'war'
-
-
+apply from: "${project.projectDir}/../../gradle/publish-java.gradle"
+apply from: "${project.projectDir}/../../gradle/warnings.gradle"
 
 dependencies {
-  compile(platform(project(':boms:geode-all-bom')))
-  implementation('javax.servlet:javax.servlet-api')
-  implementation(project(':geode-deployment:geode-deployment-legacy'))
-}
+    implementation(platform(project(':boms:geode-all-bom')))
+    compileOnly(platform(project(':boms:geode-all-bom')))
 
-war {
-    version = ''
+    implementation(project(':geode-core'))
+    implementation(project(':geode-common'))
+    implementation('io.swagger:swagger-annotations')
+
+    compileOnly(project(':geode-logging'))
+    compileOnly('com.fasterxml.jackson.core:jackson-databind')
+
+    testImplementation(project(':geode-junit'))
+
+    integrationTestImplementation(project(':geode-junit'))
 }
