+++ geode-deployment/geode-deployment-legacy/src/main/resources/META-INF/services/org.apache.geode.deployment.internal.JarDeploymentService
@@ -1,0 +1,1 @@
+org.apache.geode.deployment.internal.LegacyJarDeploymentService
