+++ settings.gradle
@@ -69,7 +69,8 @@ include 'geode-concurrency-test'
 include 'boms:geode-client-bom'
 include 'boms:geode-all-bom'
 include 'static-analysis:pmd-rules'
-include 'geode-deployment-legacy'
+include 'geode-deployment'
+include 'geode-deployment:geode-deployment-legacy'
 
 
 ['1.0.0-incubating',
